Explorer in green, dark mode
There was in Gondolin a jewel-smith named Enerdhil, the greatest of that craft among the Noldor after the death of Fëanor. Enerdhil loved all green things that grew, and his greatest joy was to see the sunlight through the leaves of trees. And it came into his heart to make a jewel within which the clear light of the sun should be imprisoned, but the jewel should be green as leaves. And he made this thing, and even the Noldor marvelled at it. For it is said that those who looked through this stone saw things that were withered or burned healed again or as they were in the grace of their youth, and that the hands of one who held it brought to all that they touched healing from hurt. This gem Enerdhil gave to Idril the King's daughter, and she wore it upon her breast and so it was saved from the burning of Gondolin. And be-fore Idril set sail she said to Eärendil her son: "The Elessar I leave with thee, for there are grievous hurts to Middle-earth which thou maybe shalt heal. 

This "tuning" is entirely based on the original Orthanc Explorer that can be found in https://www.orthanc-server.com/

Motivated to avoid the visual stress of a white background, the interface matches the colors of the Osimis Webviewer Elesser adaption.

Pre-requisites

The login/register are running in traditional PHP. The translation has been done using  gettext 

The JPG and PDF uploaders rely on CURL, DCMTK and GhostScript 

Going Online

To go online the files must be placed in a web served folder (such as /var/www/html) 

Placing the files under a "xpl" folder and grant www-data access to the tree is recommended.

See the suggested nginx configuration file.